AppServiceProvider optimized & conversion chat_id to int logger handler

+++ app/Logging/Telegram/TelegramLoggerHandler.php
@@ -18,7 +18,7 @@ class TelegramLoggerHandler extends AbstractProcessingHandler
     {
         TelegramBotApi::sendMessage(
             config('logging.channels.telegram.token'),
-            config('logging.channels.telegram.chat_id'),
+            (int)config('logging.channels.telegram.chat_id'),
             "{$record['message']}
             \n{$record['datetime']->format('Y-m-d H:i:s')}"
         );
